+++ src/main/java/dk/dataforsyningen/arkivmeta/protokol/apimapper/ProtokolDtoMapper.java
@@ -145,7 +145,7 @@ public class ProtokolDtoMapper implements RowMapper<ProtokolDto> {
         dto.setHerredsnavn(rs.getString("herredsnavn"));
         dto.setProtokoltype(rs.getString("protokoltype"));
         dto.setSognenavn(rs.getString("sognenavn"));
-        dto.setSogneid(rs.getInt("sogneid"));
+        dto.setSogneid(rs.getBigDecimal("sogneid"));
 
         return dto;
     }
